feat: add a script to convert from layerzero wire exported transactions to safe
`lz:oapp:wire` is a hardhat command that generates a series of transactions to configure the cross chain parameters for a layerzero messaging application.

unfortunately this command, while it does have an option to send transactions to Safe, either this capability is not currently working or its extremely unclear/undocumented how to use it. In either case, there is another option `--output-filename` which allows for the list of configuration transactions to be exported.
once this list is obtained, its possible to use this script to split it by network and load it into the Safe UI using the Safe Transaction Builder JSON Import feature.

this script can be tested by using the sample wire output (to be attached)