feat: script to convert from layerzero wire exported transactions to safe format (#3)
`lz:oapp:wire` is a hardhat command that generates a series of
transactions to configure the cross chain parameters for a layerzero
messaging application. It is necessary for deployment of a LayerZero app
(short of manually configurating transactions, anyway...)

unfortunately this command, while it does have an option to send
transactions to Safe using commandline option `--safe`, either this
capability is not currently working or its extremely
unclear/undocumented how to use it. In either case, there is another
option `--output-filename` which allows for the list of configuration
transactions to be exported. once this list is obtained, its possible to
use this script to split it by network and load it into the Safe UI
using the Safe Transaction Builder JSON Import feature.

this script can be tested by using the sample wire output
[txns.json](https://github.com/user-attachments/files/22621182/txns.json).
and then execute the command from repo root:

```
node ./scripts/convert-to-safe-format.js txns.json
```

It should output 3 files which can then be imported into any gnosis safe
using the Safe Transaction Builder JSON import (in the case of the
testing file, `0x5002Ee88591876C148443cdED7a6D2e466B0147a` for mainnet,
bsc, and avax).