feat: base changes from the template for the COW token (#8)
## Description

This makes the minimum necessary changes to CoWify the contracts in the
repo

## Context

The base of the repo from this commit is what we get from running the
raw layerzero project generation command. However, these contracts are
extremelye bland and do not include some of the changes we want to
support on our chains for the CoW token.

Specific changes made:
* add ERC20Permit extension to the COW token
* rename files to COW for officially
* explicitly set the name of the CoW token since it should always be the
same
* enable viaIR since it doesn't seem to be practically possible to set
all the constructor params without it

## Out of Scope
This is the minimum changes necessary to CowIfy. Tests are not complete
in this PR, and auxillery deployments.

## Testing Instructions
Tests are not yet funcxtional in this PR. see #9