feat: add additional tests (#9)
## Description
Add additional tests which would be missing for the OFT token itself
(not OFTAdapter)

## Context
These tests were pulled over from the generated repo from `npx
create-lz-app --example oft` without selecting the OFTAdapter when
prompted. Its unclear why the original template did not include these
files in the first place, but for completeness we should be running
tests against the OFT token as well.

## Out of Scope
Does not cover any deployments or any changes relating to the contracts
themselves.

## Testing Instructions

There are both forge and hardhat tests.

For the forge tests,  run `forge test` in the root of the repo.
For the hardhat tests, run `pnpm i && pnpm hardhat test` in the root of
the repo (with [pnpm](https://pnpm.io/installation) installed if you
havent already)