feat: add initial deployments (#5)
## Description
Adds the initial deployments 
 
## Context

Now that we have completed development of the lz-token, we are deploying
it initially to Mainnet, BSC, and Avalanche, the 3 chains that currently
lack the token. This PR contains the output JSON deployment artifacts
having deployed the actual tokens to all of these networks.

As a note, these deployment artfiacts are a prerequisite for running the
LayerZero wiring commands, which automatically consume these deployment
to determine the addresses and any other metadata information.

The deployment was completed by following the instructions in the README
of this repo, and wiring is pending confirmation of transactions
currently staged to the safe (see below).

## Out of scope
The contracts themselves, anything that doesnt have to do with verifying
that the expected contracts were not actually put on chain correctly.

## How to Validate

Probably the easiest way is to compare the bytecode on-chain with the
bytecode generated by solidity after running `pnpm compile`. For
example, for the Mainnet OftAdapter:

```bash
# compile so we have the correct artifacts locally
pnpm compile
# get mainnet code for the deployed contract 
cast code 0x5c7E91F4a8e58e701bFE621990bd6Fa4FD7c5379 --rpc-url $RPC_URL_1
# get bytecode generated by solidity compilation
cat artifacts/contracts/CowOftAdapter.sol/CowOftAdapter.json | jq .deployedBytecode -r
```

On my end I paste the outputs of these two commands into an editor
without word wrap, and I can see:
* the strings are the same length
* the code appears to be all the same--**excluding a couple places where
the immutable variable is injected**

Additionally, the conrtcats have been verified on all relevant networks.
The input arguments for `owner` and `lz_endpoint` can be verified by
checking the explorer sites.

## Other Relevant Info

Final wiring is still in progress and has been staged to the dev
multisig on all networks:
* mainnet
https://app.safe.global/transactions/tx?safe=eth:0x423cEc87f19F0778f549846e0801ee267a917935&id=multisig_0x423cEc87f19F0778f549846e0801ee267a917935_0xf8644e99ececb4143a258b035ce2107582ed882402c27fb0c62e906b28f36e30
* bsc
https://app.safe.global/transactions/tx?safe=bnb:0x423cEc87f19F0778f549846e0801ee267a917935&id=multisig_0x423cEc87f19F0778f549846e0801ee267a917935_0xbbda0006515e5af0d696354f049c733b655f7dfae6d5b085d32cea48d5c9967b
* avalanche
https://app.safe.global/transactions/tx?safe=avax:0x423cEc87f19F0778f549846e0801ee267a917935&id=multisig_0x423cEc87f19F0778f549846e0801ee267a917935_0xa273de2f6f0e95cff80647f8eafc9211818fd8bd2992ffd85cb0a3f318b92936