this sets up the layerzero configuration file to what we actually need (#12)
## Description
Sets the LayeZero configuration required in order to complete
deployments.

## Context
We forgot to include this file, and so the `pnpm lz:oapp:wire` command
is not working as expected.

## Out of Scope
This is just setting configuration. Running `pnpm lz:oapp:wire` should
be sufficient for identifying if things are working as expected or not.

## Testing Instructions
See the deployments section in the [README](README.md).

---------

Co-authored-by: carlos-cow <carlos@cow.fi>