fix readme again (#13)
## Description
Update the README to more like what we would expect for COW

## Context
An initial comprehensive pass of the readme was done in the `fix/readme`
branch from a few weeks before this PR. However, some of the changes in
this branch seem to have been lost. This restores lost changes from this
branch, and then merges in any updates that emerged since that time.

## Testing Instructions
To prevent issues this time, its probably best to read the `README.md`
outside of the actual diff (aka, by clicking "View File" and reading the
file from beginning to end directly).

---------

Co-authored-by: Federico Giacon <58218759+fedgiac@users.noreply.github.com>